rename documentation files accordingly
