+++ docs/scripts/add_network.md
