+++ docs/scripts/add_network_from_txt.md
