+++ docs/scripts/get_hw_id.md
