+++ docs/scripts/get_ip.md
