+++ docs/scripts/get_mac.md
