+++ docs/scripts/get_ssid.md
