+++ docs/scripts/setup_home_network.md
