+++ src/scripts/add_network_from_txt.py
