+++ src/scripts/get_hw_id.py
