+++ src/scripts/setup_home_network.py
@@ -5,8 +5,8 @@ import os
 import subprocess
 import sys
 
-import CSM_wpa_country
 import stdiomask
+import wpa_country
 
 
 class HiddenPrints:
@@ -95,9 +95,9 @@ def set_home_network():
 
     try:
         with HiddenPrints():
-            CSM_wpa_country.get_country("/etc/wpa_supplicant/wpa_supplicant.conf")
+            wpa_country.get_country("/etc/wpa_supplicant/wpa_supplicant.conf")
     except SystemExit:
-        CSM_wpa_country.update_country("/etc/wpa_supplicant/wpa_supplicant.conf", "US")
+        wpa_country.update_country("/etc/wpa_supplicant/wpa_supplicant.conf", "US")
     if is_there_password:
         items.append("-p")
         items.append(password)
